+++ packages/translator-tags/src/__tests__/fixtures/attr-class/__snapshots__/dom.expected/template.js
@@ -18,7 +18,7 @@ const _expr_Text_c_d = /* @__PURE__ */_intersection(3, _scope => {
         b: c,
         d
       }],
-      renderBody() {
+      renderBody: () => {
         _write("Hello");
       }
     }