+++ packages/translator-tags/src/core/__flush_here_and_after__.ts
@@ -1,13 +1,17 @@
-import type { Tag } from "@marko/babel-utils";
-import { currentProgramPath } from "../visitors/program";
+import { getTagDef, type Tag } from "@marko/babel-utils";
 
 export default {
-  migrate: [
+  transform: [
     (tag) => {
       // TODO: this is kinda just "flush here" right now
       // but it should replicate the behavior of the tag in Marko 5
+
+      if (getTagDef(tag)?.renderer) {
+        // we're in the compat layer and this is being handled by Marko 5.
+        return;
+      }
+
       tag.replaceWithMultiple(tag.node.body.body);
-      currentProgramPath.scope.crawl();
     },
   ],
 } as Tag;