+++ packages/translator-tags/src/core/id.ts
@@ -53,4 +53,5 @@ export default {
       descriptionMoreURL: "https://markojs.com/docs/core-tags/#id",
     },
   ],
+  types: "@marko/translator-tags/tag-types/id.d.marko",
 } as Tag;