+++ packages/marko/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 5.33.12
+
+### Patch Changes
+
+- [#2187](https://github.com/marko-js/marko/pull/2187) [`fe344b0`](https://github.com/marko-js/marko/commit/fe344b00041677f40ce49d03c0fb283322a1c898) Thanks [@DylanPiercey](https://github.com/DylanPiercey)! - Fix some tags/class interop issues.
+
+- Updated dependencies [[`fe344b0`](https://github.com/marko-js/marko/commit/fe344b00041677f40ce49d03c0fb283322a1c898)]:
+  - @marko/compiler@5.35.10
+  - @marko/translator-default@5.32.11
+
+## 5.33.11
+
+### Patch Changes
+
+- [#2182](https://github.com/marko-js/marko/pull/2182) [`1fa3b05`](https://github.com/marko-js/marko/commit/1fa3b056006d5d0e3ac221b5b4a18b78de5add21) Thanks [@DylanPiercey](https://github.com/DylanPiercey)! - Support tracking the "input" using babels scope analysis.
+
+- Updated dependencies [[`1fa3b05`](https://github.com/marko-js/marko/commit/1fa3b056006d5d0e3ac221b5b4a18b78de5add21)]:
+  - @marko/compiler@5.35.9
+  - @marko/translator-default@5.32.10
+
 ## 5.33.10
 
 ### Patch Changes